requirements.txt
streamlit==1.39.0
pandas==2.2.3
PyGithub==2.4.0